Move argcomplete code to conda_argparse
This way, subcommands won't need to enable it manually. Just using the conda
ArgumentParser will be enough.
